ECS::ComponentArray worked

+++ Source/Engine/Source/ING/ECS/Component/Component.h
@@ -20,7 +20,7 @@ namespace ING {
 
 	namespace ECS {
 
-		using ComponentId = unsigned long;
+		using ComponentId = unsigned int;
